+++ demos/04-vectorfield/index.js
@@ -49,10 +49,18 @@
 	    );
 	    pb.config.preDraw = function() { preDraw(); };
 
+	    // A list of vectors (position plus velocity)
 	    var pointlist = [];
-	    for( var i = 0; i < 100; i++ ) {
-		pointlist.push( Vertex.randomVertex(pb.viewport()) );
+	    
+	    var resetPointCloud = function() {
+		pointlist = [];
+		var v;
+		for( var i = 0; i < 100; i++ ) {
+		    v = Vertex.randomVertex(pb.viewport());
+		    pointlist.push( new Vector(v, v.clone()) );
+		}
 	    }
+	    resetPointCloud();
 	    
 
 	    // +---------------------------------------------------------------------------------
@@ -73,7 +81,8 @@
 	    // +-------------------------------
 	    var config = PlotBoilerplate.utils.safeMergeByKeys( {
 		showFieldVectors      : true,
-		animate               : false
+		animate               : false,
+		resetPointCloud       : function() { resetPointCloud(); }
 	    }, GUP );
 	    
 
@@ -89,7 +98,6 @@
 	    pb.add( cv3 = new Vector( new Vertex(controlOffset,-controlOffset), new Vertex(controlOffset+diameter*0.6,-controlOffset-diameter*0.6) ), false );
 	    pb.add( cv4 = new Vector( new Vertex(controlOffset,controlOffset), new Vertex(controlOffset+diameter*0.6,controlOffset+diameter*0.6) ), false );
 	    var controlVectors = [ cv0, cv1, cv2, cv3, cv4 ];
-	    //var field = [];
 	    
 	    
 	    // +--------------------------------------------------------------------------------
@@ -113,9 +121,7 @@
 		    pb.draw.dot( vector.a, '#e800e8' );
 	    }
 	    function preDraw() {
-		//console.log( 'preDraw' );		    
 		if( config.showFieldVectors ) {
-		    //console.log( 'draw field vectors' );
 		    for( var x = 0; x < pb.canvasSize.width/2; x+=displayStep ) {
 			for( var y = 0; y < pb.canvasSize.height/2; y+=displayStep ) {
 			    drawFieldVectorAt( vert(x,y) );
@@ -126,70 +132,33 @@
 		    }
 		}
 		if( config.animate ) {
-		    //console.log('draw point cloud');
 		    for( var i in pointlist ) {
-			pb.fill.circle( pointlist[i], 2, 'black' );
+			pb.draw.arrow( pointlist[i].a, pointlist[i].b, 'black' );
+			pb.fill.circle( pointlist[i].a, 2, 'black' );
 		    }
 		}
 		    
 	    };
 
-	    /*
-	    function adjustVectorField() {
-		 for( var fi in vectorField ) {
-		     var anyVector = vectorField[fi];
-		     // Set to length 0
-		     anyVector.b.set(anyVector.a);
-		     
-		     for( var ci in controlVectors ) {
-			 var controlVector = controlVectors[ci];
-			 var strength = controlVector.length();
-			 adjustVector_radial( controlVector, strength, anyVector );
+	    function adjustPointList( stepFactor ) {
+		for( var fi in pointlist ) {
+		    var pointWithVelocity = pointlist[fi];
+		    var anyVector = pointWithVelocity.clone();
+		    anyVector.b.set( anyVector.a );
+		    
+		    for( var ci in controlVectors ) {
+			var controlVector = controlVectors[ci];
+			var strength = controlVector.length();
+			adjustVector_radial( controlVector, strength, anyVector );
 
-		     }
+		    }
+
+		    // Relocate vector to a=(0,0), but keep an only copy.
+		    pointWithVelocity.a.add( anyVector.clone().sub(anyVector.a).b.multiplyScalar(stepFactor) );
+		    pointWithVelocity.b.set( anyVector.b );
 		}
 		pb.redraw();
 	    }
-	    */
-
-	     function adjustPointList( stepFactor ) {
-		 for( var fi in pointlist ) {
-		     var point = pointlist[fi];
-		     var anyVector = new Vector( point.clone(), point.clone() );
-		     
-		     for( var ci in controlVectors ) {
-			 var controlVector = controlVectors[ci];
-			 var strength = controlVector.length();
-			 adjustVector_radial( controlVector, strength, anyVector );
-
-		     }
-
-		     // Relocate vector to a=(0,0)
-		     anyVector.sub( anyVector.a );
-		     pointlist[fi].add( anyVector.b.multiplyScalar(stepFactor) );
-		}
-		pb.redraw();
-	    }
-
-	    /*
-	      // Some test which still looks nice
-	    function adjustVector_radial( controlVector, strength, anyVector ) {
-		var t = controlVector.getClosestT(anyVector.a);
-		// t = Math.max(0, Math.min(1,t));
-		var pointOnControlVector = controlVector.vertAt(t);
-		var dist = pointOnControlVector.distance( anyVector.a );
-		// Is the vertex inside the strengh area of thi control Vector?
-		if( dist < strength ) {
-		    // console.log('x');
-
-		    // dist -= strength;
-		    anyVector.b.x += (anyVector.a.x-controlVector.a.x)*(strength/dist);
-		    anyVector.b.y += (anyVector.a.y-controlVector.a.y)*(strength/dist);
-		    //anyVector.b.x += (anyVector.a.x-controlVector.a.x)*(dist/strength);
-		    //anyVector.b.y += (anyVector.a.y-controlVector.a.y)*(dist/strength);
-		}
-	    }
-	    */
 
 	    function adjustVector_radial( controlVector, strength, anyVector ) {
 		var dist = controlVector.a.distance( anyVector.a );
@@ -201,20 +170,12 @@
 		
 	    }
 
-	    // Install listeners to all control vectors
-	    /*for( var i in controlVectors ) {
-		controlVectors[i].a.listeners.addDragListener( function(e) { adjustVectorField(); } );
-		controlVectors[i].b.listeners.addDragListener( function(e) { adjustVectorField(); } );
-	    }
-	    adjustVectorField();
-	    */
-
-
 	    function renderAnimation() {
 		//adjustVectorField();
 		adjustPointList( 0.01 );
 		//pb.redraw();
-		window.requestAnimationFrame( renderAnimation );
+		if( config.animate )
+		    window.requestAnimationFrame( renderAnimation );
 	    };
 	    
 	    function toggleAnimation() {
@@ -233,6 +194,7 @@
 		var f0 = gui.addFolder('Points');
 		f0.add(config, 'showFieldVectors').onChange( function() { pb.redraw(); } ).name('Show field vectors').title('Show field vectors.');
 		f0.add(config, 'animate').onChange( function() { toggleAnimation(); } ).name('Animate a point cloud').title('Animate a point cloud.');
+		f0.add(config, 'resetPointCloud').name('Reset point cloud').title('Reset point cloud.');
 		f0.open();
 	    }