+++ src/Vector.js
@@ -10,7 +10,8 @@
  * @date     2019-01-30
  * @modified 2019-02-23 Added the toSVGString function, overriding Line.toSVGString.
  * @modified 2019-03-20 Added JSDoc tags.
- * @version  1.0.2
+ * @modified 2019-04-19 Added the clone function (overriding Line.clone()).
+ * @version  1.0.3
  *
  * @file Vector
  * @public
@@ -32,6 +33,20 @@
 	Line.call(this,vertA,vertB);
     };
     Object.extendClass(Line,Vector);
+
+
+    /**
+     * Create a deep clone of this Vector.
+     *
+     * @method clone
+     * @override
+     * @return {Vector} A copy if this line.
+     * @instance
+     * @memberof Vector
+     **/
+    Line.prototype.clone = function() {
+	return new Vector( this.a.clone(), this.b.clone() );
+    };
 
     
     /**